requirements.txt
torch>=2.1.0
torchvision>=0.16.0
accelerate>=0.27.0
transformers>=4.46.0
sentencepiece>=0.2.0
timm>=0.9.16
einops>=0.7.0
numpy>=1.26.0
scipy>=1.11.0

opencv-python-headless>=4.9.0
pillow>=10.3.0

pandas>=2.2.0
typing_extensions>=4.10.0
xmltodict>=0.13.0
babel>=2.15.0
python-dotenv>=1.0.1

huggingface_hub>=0.23.0
hf_xet

streamlit>=1.33.0
jupyterlab>=4.1.0

langchain-google-genai==3.0.1